+++ deadchest-core/src/main/java/me/crylonz/deadchest/DeadChestLoader.java
@@ -260,6 +260,7 @@ public class DeadChestLoader {
         config.register(ConfigKey.STORE_XP_PERCENTAGE.toString(), 100);
         config.register(ConfigKey.KEEP_INVENTORY_ON_PVP_DEATH.toString(), false);
         config.register(ConfigKey.KEEP_INVENTORY_ON_PVP_WORLDS.toString(), Collections.emptyList());
+        config.register(ConfigKey.VANILLA_DROP_WORLDS.toString(), Collections.emptyList());
         config.register(ConfigKey.LOCALIZATION_LANGUAGE.toString(), "en");
         config.register(ConfigKey.REPLACE_OLDEST.toString(), false);
         config.register(ConfigKey.PLACEMENT_SAFE_LOCATION.toString(), true);